Rename generateEqualityOperators.ts to generateOperators.ts
