+++ src/generateOperators.ts
